trial_1/server/requirements.txt
absolufy-imports==0.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
audioop-lts==0.2.2
Authlib==1.6.3
azure-ai-documentintelligence==1.0.2
azure-core==1.35.0
azure-identity==1.24.0
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
cobble==0.1.4
coloredlogs==15.0.1
cryptography==45.0.7
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
fastapi==0.116.1
filelock==3.19.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.9.0
google-adk==1.14.1
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.114.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.37.0
google-cloud-bigtable==2.32.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.57.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.36.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
humanize==4.13.0
idna==3.10
importlib_metadata==8.7.0
instructor==1.11.3
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.1
magika==0.6.2
Mako==1.3.10
mammoth==1.10.0
markdown-it-py==4.0.0
markdownify==1.2.0
markitdown==0.1.3
MarkupSafe==3.0.2
mcp==1.14.0
mdurl==0.1.2
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
networkx==3.5
numpy==2.3.3
olefile==0.47
onnxruntime==1.22.1
openai==1.107.1
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
pandas==2.3.2
pdfminer.six==20250506
pglast==7.2
pgvector==0.4.1
pillow==11.3.0
postgres-mcp==0.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
sqlalchemy-spanner==1.16.0
sqlparse==0.5.3
sse-starlette==3.0.2
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.47.3
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.22.0
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
websockets==15.0.1
xlrd==2.0.2
xlsxwriter==3.2.5
yarl==1.20.1
youtube-transcript-api==1.0.3
zipp==3.23.0
